requirements.txt
flask
Flask-SQLAlchemy
flask-login
better_profanity
mysqlclient
mysql-connector-python
folium
geocoder

